% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic12stars.R
\name{magic12stars}
\alias{magic12stars}
\title{Magic 12 Stars

This function asks for an input of a date and
returns the corresponding astrological sign
if one is born on that day.

The function uses the lubridate() package to
get the year, month, and the day from a given date.
The code is from Vitalie Spinu (github@vspinu)}
\arguments{
\item{testdate}{What is the chosen date to be tested? Defaults to the birthday of this package.}
}
\description{
Magic 12 Stars

This function asks for an input of a date and
returns the corresponding astrological sign
if one is born on that day.

The function uses the lubridate() package to
get the year, month, and the day from a given date.
The code is from Vitalie Spinu (github@vspinu)
}
\examples{
magic12stars()
}
